Changed return type of methods to double so they can be rounded

+++ SnN Health Monitor - Drive/BytesToBinaryUnits.cs
@@ -13,12 +13,12 @@ namespace Drive
 
         // this list gets indexed by an enum as some values are too large for it to hold.
 
-        public static float Convert(long inputBytes, BinaryUnitsCode unitOut)
+        public static double Convert(long inputBytes, BinaryUnitsCode unitOut)
         {
             if (inputBytes <= 0)
                 throw new ArgumentOutOfRangeException("The inputed Bytes need to be larger than 0 in order to convert them");
 
-            return (float) inputBytes / _unitsArray[(byte)unitOut];
+            return (double) inputBytes / _unitsArray[(byte)unitOut];
         }
 
         public static float GetUnitInBytes(BinaryUnitsCode unit)
